requirements.txt
Flask==0.12.2
cloudant==2.4.0
mysql-connector-python== 8.0.15